added required xp for next level function to be just simple cube function for now, also changed everything related to xp to be a ulong.

+++ Assets/scripts/Character Scripts/Enemy.cs
@@ -5,11 +5,11 @@ using System.Collections.Generic;
 
  public class Enemy : Actor { 
 
-	private int baseXp; //unmodified xp granted when defeated
+	private ulong baseXp; //unmodified xp granted when defeated
 	private  Dictionary<GameItem, int> droppableItems;  //list of items that can be dropped from the enemy. along with corresponding probability of being dropped.
 
 //	protected int levelOffset; //Determines the offset of the enemy's level to the player
-	public Enemy(List<CombatAbility> combatMoveSet, string characterName, int level, int healthCap, Gender gender, string id, int baseXp) : 
+	public Enemy(List<CombatAbility> combatMoveSet, string characterName, int level, int healthCap, Gender gender, string id, ulong baseXp) : 
 	base(combatMoveSet, characterName, level, healthCap, gender, id) {
 
 		droppableItems = new Dictionary<GameItem, int> ();
@@ -26,11 +26,11 @@ using System.Collections.Generic;
 		//Add droppable item with a specified drop chance
 	}
 
-	public int getGrantableXp() {
+	public ulong getGrantableXp() {
 		/*takes base xp, multiplies it by level, and in future modifies based on player level.
 		 */
 
-		return baseXp * level;
+		return baseXp * (ulong)level;
 	}
